chatbot.md
This a chatbot project just like the What's App where you can see on this notification  bar that who sent you the message and what message has been sent by you to the other person.
All you have to do is  open XAMPP and in the folder of htdocs create a folder of your name and in that folder three more  folder are to be created: includes, public, scripts and c.jpg and index.php. In the 'inlcludes' folder two subfolders have to be created: database("connection.php"), function("chat_fun.php","connection.php") and a core.inc file.
In the folder 'public' a subfolder of name 'css' has to be created which contains "main.css".
And In the folder 'scripts'  two subfolders have to created: 'js'("auto_chat.js", "jquery-3.3.1.js", "send.js") and 'php'("chat.php", "send.php").
All the files mentioned above are uploaded,a ll you have to do is just follow the instructions.